requirements.txt
requests
python-telegram-bot
schedule
Pillow
fpdf2
urllib3
beautifulsoup4
deep-translator